requirements.txt
pip3
Jinja2
psycopg2-binary
flask